requirements.txt
torch>=2.4.0
torchvision>=0.19.0
transformers>=4.45.0
safetensors>=0.4.0
diffusers>=0.30.0
flash-attn>=2.5.0
Pillow>=10.0.0
numpy>=1.24.0
tqdm>=4.65.0
accelerate>=0.26.0
torchdiffeq